refactor(packages): move fragments/activity into view
