+++ app/src/main/res/layout/fragment_camera.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CameraFragment">
+    tools:context=".viewmodel.CameraFragment">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/viewFinder"
